add some basic retry logic to rabbitmq code general tidy of rabbitmq work
