+++ bundlegen/rabbitmq/example.env
@@ -1,11 +1,11 @@
 # Where BundleGen will store the generated bundles
 BUNDLE_STORE_DIR=/home/vagrant/rabbitmq_testing/bundles/
 # URL to the rabbitmq broker
-RABBITMQ_BROKER=localhost
+RABBITMQ_HOST=localhost
 # Port RabbitMQ is listening on (if not set, will default to 5672)
 RABBITMQ_PORT=1234
 # Temporary working directory for BundleGen (images will be downloaded here)
-TMP_DIR=/tmp/bundlegen
+BUNDLEGEN_TMP_DIR=/tmp/bundlegen
 # username and password to access the OCI registry
 RDK_OCI_REGISTRY_CREDS=username:password
 # Where to search for platform template json files (recursive search)
