+++ bundlegen/rabbitmq/result.py
@@ -17,7 +17,8 @@
 
 from enum import Enum
 
+
 class Result(int, Enum):
     SUCCESS = 0
-    TRANSIENT_ERROR = 1,
+    TRANSIENT_ERROR = 1
     FATAL_ERROR = 2
