+++ docker/rabbitmq/docker-compose.yml
@@ -5,7 +5,7 @@ services:
     image: bundlegen-rabbit:latest
     environment:
         - BUNDLE_STORE_DIR=/bundles
-        - RABBITMQ_BROKER=bundlegen-rabbit
+        - RABBITMQ_HOST=bundlegen-rabbit
         - TMP_DIR=/tmp/
     volumes:
         - generated_bundles:/bundles
